* commit '58236b6874c70063bba65dc79cc42268c6cfaa7a':
  media: verify secure, adaptive and flexYUV video capabilities

+++ tests/tests/media/src/android/media/cts/VideoEncoderTest.java
@@ -229,7 +229,7 @@ public class VideoEncoderTest extends MediaPlayerTestBase {
                 if (bitRate <= 0) {
                     bitRate = encCaps.getBitrateRange().clamp(
                         (int)(encCaps.getBitrateRange().getUpper() /
-                                Math.sqrt(maxWidth * maxHeight / width / height)));
+                                Math.sqrt((double)maxWidth * maxHeight / width / height)));
                 }
                 outFmt.setInteger(MediaFormat.KEY_BIT_RATE, bitRate);
